Clean up * remove unneccesary dependencies * move scxml.py to tests (since its just utils for tests) * remove leftover files
