--- package.py
@@ -1,10 +1,0 @@
-def convert(my_name):
-    """
-    Print a line about converting a notebook.
-    Args:
-        my_name (str): person's name
-    Returns:
-        None
-    """
-
-    print(f"I'll convert a notebook for you some day, {my_name}.")
