+++ setup.cfg
@@ -18,8 +18,3 @@ classifiers =
 zip_safe = False
 include_package_data = True
 packages = xstate
-install_requires =
-    ipython>=6
-    nbformat>=4
-    nbconvert>=5
-    requests>=2
