+++ tests/test_scxml.py
@@ -2,7 +2,7 @@ import xml.etree.ElementTree as ET
 from typing import Optional, Dict, List
 import json
 import pytest
-from xstate.scxml import scxml_to_machine
+from tests.scxml import scxml_to_machine
 from pprint import PrettyPrinter
 
 pp = PrettyPrinter(indent=2)
